added assignment of courses by admin

+++ lib/models/charts/bar_charts/custom_bar_chart_data.dart
@@ -1,6 +1,6 @@
 class CustomBarChartData {
   final String x;
-  final double y;
+  final int y;
 
   CustomBarChartData({required this.x, required this.y});
 }
